+++ backend/canisters/storage_index/impl/src/updates/c2c_vault_ops.rs
@@ -64,6 +64,7 @@ fn c2c_vault_ops_impl(args: Args, state: &mut RuntimeState) -> Response {
                     bucket_vault::VaultOp::SetLegalHold(bucket_vault::SetLegalHoldOp {
                         file_id: l.blob_reference.blob_id,
                         legal_hold: l.legal_hold,
+                        reference: l.reference,
                     }),
                 );
             }